backend/weather-advice/requirements.txt
google-cloud-firestore
google-cloud-secret-manager
requests
google-genai